Gallery Update - Background etc.
Background and sidebar added.  Will put the Highslide in a div once the
real thing is up and not just the hardcoded version, and will replace
"print" button with an image like the ones on the canvas, as well as
switch it so it does not print the whole page, just the card image.